Don't wipe request when converting to Response.
Here's generated dispatch code:

return static_cast<S*>(this)->setBinaryState(
	req.getArg<bool>(Arg::BinaryState),
	req); ///< This creates ActionResponse object from ActionRequest

C++ does **not** guarantee that method parameters are processed left -> right.
This works on actual hardware, but when run in Linux Host mode it fails because
the input parameter is read **after** the construction of ActionResponse.

We now just set a flag inside ActionResponse and don't wipe the request until actually necessary.
The alternative is more verbose wrapper code.